chore: moving some files around and exporting things as alpha services
