+++ packages/backend-test-utils/report-alpha.api.md
@@ -1,0 +1,85 @@
+## API Report File for "@backstage/backend-test-utils"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { ActionsRegistryActionOptions } from '@backstage/backend-plugin-api/alpha';
+import { ActionsRegistryService } from '@backstage/backend-plugin-api/alpha';
+import { ActionsService } from '@backstage/backend-plugin-api/alpha';
+import { ActionsServiceAction } from '@backstage/backend-plugin-api/alpha';
+import { AnyZodObject } from 'zod';
+import { BackstageCredentials } from '@backstage/backend-plugin-api';
+import { JsonObject } from '@backstage/types';
+import { JsonValue } from '@backstage/types';
+import { LoggerService } from '@backstage/backend-plugin-api';
+import { ServiceFactory } from '@backstage/backend-plugin-api';
+
+// @alpha (undocumented)
+export function actionsRegistryServiceMock(options?: {
+  logger: LoggerService;
+}): MockActionsRegistry;
+
+// @alpha (undocumented)
+export namespace actionsRegistryServiceMock {
+  const // (undocumented)
+    factory: () => ServiceFactory<
+      ActionsRegistryService,
+      'plugin',
+      'singleton'
+    >;
+  const // (undocumented)
+    mock: (
+      partialImpl?: Partial<ActionsRegistryService> | undefined,
+    ) => ServiceMock<ActionsRegistryService>;
+}
+
+// @alpha (undocumented)
+export namespace actionsServiceMock {
+  const // (undocumented)
+    factory: () => ServiceFactory<ActionsService, 'plugin', 'singleton'>;
+  const // (undocumented)
+    mock: (
+      partialImpl?: Partial<ActionsService> | undefined,
+    ) => ServiceMock<ActionsService>;
+}
+
+// @alpha
+export class MockActionsRegistry
+  implements ActionsRegistryService, ActionsService
+{
+  // (undocumented)
+  readonly actions: Map<string, ActionsRegistryActionOptions<any, any>>;
+  // (undocumented)
+  static create(opts: { logger: LoggerService }): MockActionsRegistry;
+  // (undocumented)
+  invoke(opts: {
+    id: string;
+    input?: JsonObject;
+    credentials?: BackstageCredentials;
+  }): Promise<{
+    output: JsonValue;
+  }>;
+  // (undocumented)
+  list(): Promise<{
+    actions: ActionsServiceAction[];
+  }>;
+  // (undocumented)
+  register<
+    TInputSchema extends AnyZodObject,
+    TOutputSchema extends AnyZodObject,
+  >(options: ActionsRegistryActionOptions<TInputSchema, TOutputSchema>): void;
+}
+
+// @alpha (undocumented)
+export type ServiceMock<TService> = {
+  factory: ServiceFactory<TService>;
+} & {
+  [Key in keyof TService]: TService[Key] extends (
+    ...args: infer Args
+  ) => infer Return
+    ? TService[Key] & jest.MockInstance<Return, Args>
+    : TService[Key];
+};
+
+// (No @packageDocumentation comment for this package)
+```
