+++ packages/backend-test-utils/report.api.md
@@ -3,11 +3,6 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
-import { ActionsRegistryActionOptions } from '@backstage/backend-plugin-api';
-import { ActionsRegistryService } from '@backstage/backend-plugin-api';
-import { ActionsService } from '@backstage/backend-plugin-api';
-import { ActionsServiceAction } from '@backstage/backend-plugin-api';
-import { AnyZodObject } from 'zod';
 import { AuditorService } from '@backstage/backend-plugin-api';
 import { AuthorizeResult } from '@backstage/plugin-permission-common';
 import { AuthService } from '@backstage/backend-plugin-api';
@@ -29,7 +24,6 @@ import { ExtensionPoint } from '@backstage/backend-plugin-api';
 import { HttpAuthService } from '@backstage/backend-plugin-api';
 import { HttpRouterService } from '@backstage/backend-plugin-api';
 import { JsonObject } from '@backstage/types';
-import { JsonValue } from '@backstage/types';
 import Keyv from 'keyv';
 import { Knex } from 'knex';
 import { LifecycleService } from '@backstage/backend-plugin-api';
@@ -58,33 +52,6 @@ export function createMockDirectory(
 export interface CreateMockDirectoryOptions {
   content?: MockDirectoryContent;
   mockOsTmpDir?: boolean;
-}
-
-// @public
-export class MockActionsRegistry
-  implements ActionsRegistryService, ActionsService
-{
-  // (undocumented)
-  readonly actions: Map<string, ActionsRegistryActionOptions<any, any>>;
-  // (undocumented)
-  static create(opts: { logger: LoggerService }): MockActionsRegistry;
-  // (undocumented)
-  invoke(opts: {
-    id: string;
-    input?: JsonObject;
-    credentials?: BackstageCredentials;
-  }): Promise<{
-    output: JsonValue;
-  }>;
-  // (undocumented)
-  list(): Promise<{
-    actions: ActionsServiceAction[];
-  }>;
-  // (undocumented)
-  register<
-    TInputSchema extends AnyZodObject,
-    TOutputSchema extends AnyZodObject,
-  >(options: ActionsRegistryActionOptions<TInputSchema, TOutputSchema>): void;
 }
 
 // @public (undocumented)
@@ -195,32 +162,6 @@ export function mockErrorHandler(): ErrorRequestHandler<
 
 // @public
 export namespace mockServices {
-  // (undocumented)
-  export namespace actions {
-    const // (undocumented)
-      factory: () => ServiceFactory<ActionsService, 'plugin', 'singleton'>;
-    const // (undocumented)
-      mock: (
-        partialImpl?: Partial<ActionsService> | undefined,
-      ) => ServiceMock<ActionsService>;
-  }
-  // (undocumented)
-  export function actionsRegistry(options?: {
-    logger: LoggerService;
-  }): MockActionsRegistry;
-  // (undocumented)
-  export namespace actionsRegistry {
-    const // (undocumented)
-      factory: () => ServiceFactory<
-        ActionsRegistryService,
-        'plugin',
-        'singleton'
-      >;
-    const // (undocumented)
-      mock: (
-        partialImpl?: Partial<ActionsRegistryService> | undefined,
-      ) => ServiceMock<ActionsRegistryService>;
-  }
   // (undocumented)
   export namespace auditor {
     const // (undocumented)
