+++ packages/backend-test-utils/src/alpha/services/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 The Backstage Authors
+ * Copyright 2025 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,16 +14,7 @@
  * limitations under the License.
  */
 
-import { MiddlewareFactory } from '@backstage/backend-defaults/rootHttpRouter';
-import { mockServices } from '../services';
-
-/**
- * A mock for error handler middleware that can be used in router tests.
- * @public
- */
-export function mockErrorHandler() {
-  return MiddlewareFactory.create({
-    config: mockServices.rootConfig(),
-    logger: mockServices.rootLogger(),
-  }).error();
-}
+export { actionsRegistryServiceMock } from './ActionsRegistryServiceMock';
+export { MockActionsRegistry } from './MockActionsRegistry';
+export { actionsServiceMock } from './ActionsServiceMock';
+export { type ServiceMock } from './simpleMock';
