+++ packages/backend-test-utils/src/services/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Backstage Authors
+ * Copyright 2023 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,6 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-export * from './wiring';
-export * from './services';
+export { mockServices } from './mockServices';
+export { mockCredentials } from './mockCredentials';
+export { type ServiceMock } from './simpleMock';
